Fix super, separate normal_tangential_{edge|face}_transform
