+++ finat/trace.py
@@ -5,4 +5,4 @@ from finat.fiat_elements import ScalarFiatElement
 
 class HDivTrace(ScalarFiatElement):
     def __init__(self, cell, degree):
-        super(HDivTrace, self).__init__(FIAT.HDivTrace(cell, degree))
+        super().__init__(FIAT.HDivTrace(cell, degree))
